requirements.txt
matplotlib>=3.7.0
numpy>=1.24.0
Pillow>=9.5.0
pytest>=7.3.1
pytest-cov>=4.1.0 